+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/TabbedApplicationLayout.java
@@ -79,7 +79,8 @@ public class TabbedApplicationLayout implements EntityApplicationPanel.Applicati
 		applicationPanel.entityPanels().forEach(this::addTab);
 
 		//tab pane added to a base panel for correct Look&Feel rendering
-		return borderLayoutPanel(new BorderLayout())
+		return borderLayoutPanel()
+						.layout(new BorderLayout())
 						.centerComponent(tabbedPane)
 						.border(createEmptyBorder(0, Layouts.GAP.getOrThrow(), 0, Layouts.GAP.getOrThrow()))
 						.build();