+++ changelog.md
@@ -95,6 +95,7 @@ Codion Change Log
 - MenuItemBuilder, overloaded builder() methods removed.
 - ToolBarBuilder, overloaded builder() methods removed.
 - LabelBuilder, overloaded builder() methods removed.
+- BorderLayoutPanelBuilder, overloaded builder() methods removed.
 ### is.codion.swing.framework.model
 - EntityComboBoxModel.Builder.EntityTypeBuilder and ConnectionProviderBuilder added.
 ### is.codion.swing.framework.ui